refactor: split to github/slack codebase
