+++ src/github/index.ts
@@ -1,0 +1,1 @@
+export { githubApp } from './app';
