+++ src/slack/index.ts
@@ -1,0 +1,1 @@
+export { slackApp } from './app';
